:bugfix frontend for commit done
