+++ frontend/src/pages/RunningPods.vue
@@ -6,59 +6,7 @@
         Currently Running Pods
       </h1>
       <hr class="border-t-2 border-teal-400 my-6" />
-      <div
-        v-if="modal_close"
-        class="flex flex-col items-center p-2 border-2 border-teal-500 bg-white shadow-lg p-6 m-6 rounded w-fit fixed inset-0 h-auto max-w-xl mx-auto my-auto z-50"
-      >
-        <h1>New Image Details</h1>
-        <div class="flex flex-col">
-          <div class="flex items-center">
-            <label for="text-gray-900">Image Name</label>
-            <input
-              type="text"
-              v-model="container_params.image_name"
-              class="block w-full font-medium rounded-md border-1 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 m-4"
-            />
-          </div>
-          <div class="flex items-center">
-            <label for="text-gray-900">Tag</label>
-            <input
-              type="text"
-              v-model="container_params.tag"
-              class="block w-full font-medium rounded-md border-1 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 m-4"
-            />
-          </div>
-          <div class="flex items-center">
-            <label for="text-gray-900">Message</label>
-            <input
-              type="text"
-              v-model="container_params.message"
-              class="block w-full font-medium rounded-md border-1 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 m-4"
-            />
-          </div>
-          <div
-            class="flex gap-2 items-center"
-            @click="toggleActive = !toggleActive"
-          >
-            <label for="text-gray-900">Pause Container</label>
-            <div
-              class="w-16 h-10 flex items-center bg-gray-300 rounded-full p-1 duration-300 ease-in-out"
-              :class="{ 'bg-green-500': toggleActive }"
-            >
-              <div
-                class="bg-white w-8 h-8 rounded-full shadow-md transform duration-300 ease-in-out"
-                :class="{ 'translate-x-6': toggleActive }"
-              ></div>
-            </div>
-          </div>
-          <button
-            @click="commit_pod()"
-            class="bg-blue-500 text-white ml-2 mr-2 p-2 pl-4 pr-4 rounded-md hover:bg-blue-600 focus:outline focus:ring focus:border-blue-700 m-4"
-          >
-            Commit
-          </button>
-        </div>
-      </div>
+      <Modal :modal_close="modal_visible" :selected_pod="selected_pod" @close_modal="close_modal"/>
       <div class="grid grid-cols-1">
         <div
           v-for="(pod, index) in runningPods"
@@ -142,21 +90,15 @@
 
 <script>
 import SideNavbar from '@/components/SideNavbar.vue'
+import Modal from '../components/Modal.vue'
 
 export default {
   name: 'RunningPods',
   data() {
     return {
       runningPods: [],
-      modal_close: false,
-      toggleActive: false,
-      selected_pod: '',
-      container_params: {
-        image_name: '',
-        tag: '',
-        message: '',
-        pause: this.toggleActive,
-      },
+      modal_visible:false,
+      selected_pod:""
     }
   },
   mounted() {
@@ -226,59 +168,11 @@ export default {
     },
     open_modal(pod) {
       this.selected_pod = pod
-      this.modal_close = true
-    },
-    close_modal() {
-      this.modal_close = false
-    },
-    async commit_pod() {
-      try {
-        const response = await fetch(
-          '/api/method/flashdesk.api.pods.running_pods.commit_pod',
-          {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              container_id: this.selected_pod.container_id,
-              container_params: this.container_params,
-            }),
-          }
-        )
-
-        if (response.ok) {
-          const data = await response.json()
-          this.$swal({
-            toast: true,
-            position: 'bottom-right',
-            showConfirmButton: false,
-            timer: 3000,
-            icon: 'success',
-            title: 'Sucess',
-            text: 'Pod  Commited Sucessfully',
-            showCancelButton: 'true',
-          })
-          this.close_modal()
-          this.fetchRunningPods()
-        } else {
-          this.$swal({
-            toast: true,
-            position: 'bottom-right',
-            showConfirmButton: false,
-            timer: 3000,
-            icon: 'error',
-            title: 'Error',
-            text: 'Something Went Wrong with POD COMMITING :(',
-            showCancelButton: 'true',
-          })
-          console.error('Error terminating pod:', response.statusText)
-        }
-      } catch (error) {
-        console.error('Error fetching images:', error)
-      }
-    },
+      this.modal_visible = true
+    },close_modal() {
+      this.modal_visible= false
+    }
   },
-  components: { SideNavbar },
+  components: { SideNavbar, Modal },
 }
 </script>
